feat: enable all skipped tests and fix test suite
- Enable all 13 previously skipped TypeScript tests
- Fix failing tests by adding proper intent setup and fulfillment
- Update AddressConverter to validate Ethereum addresses (top 12 bytes must be zero)
- Fix test infrastructure issues including gas limit extraction and balance forwarding
- Remove untracked debug test files
- Update tests to handle new validation requirements

All 304 tests now passing (144 TypeScript + 160 Solidity)

+++ contracts/test/TestMessageBridgeProver.sol
@@ -67,7 +67,7 @@ contract TestMessageBridgeProver is MessageBridgeProver {
     // No custom events needed for testing
 
     /**
-     * @notice Mock implementation of initiateProving
+     * @notice Mock implementation of prove
      * @dev Records arguments and marks dispatched = true
      */
     function prove(
@@ -117,7 +117,7 @@ contract TestMessageBridgeProver is MessageBridgeProver {
     function addProvenIntent(bytes32 _hash, address _claimant) public {
         _provenIntents[_hash] = ProofData({
             claimant: _claimant,
-            destinationChainID: uint96(block.chainid)
+            destinationChainID: uint64(block.chainid)
         });
     }
